// src/lib/simkl/types/genre.ts
export enum GenreTypes {
	all = '',
	series = '/series',
	movies = '/movies',
	ovas = '/ovas',
	onas = '/onas'
}
export enum GenreYears {
	today = 'today',
	thisWeek = 'this-week',
	thisMonth = 'this-month',
	thisYear = 'this-year'
}
export enum GenrePopularity {
	popularToday = 'popular-today',
	popularThisWeek = 'popular-this-week',
	popularThisMonth = 'popular-this-month',
	popularThisYear = 'popular-this-year',
	rank = 'rank',
	votes = 'votes',
	releaseDate = 'release-date',
	lastAirTime = 'last-air-date',
	aToz = 'a-z',
	zToa = 'z-a'
}
export enum Genres {
	action = 'action',
	adventure = 'adventure',
	cars = 'cars',
	comedy = 'comedy',
	dementia = 'dementia',
	demons = 'demons',
	drama = 'drama',
	ecchi = 'ecchi',
	fantasy = 'fantasy',
	game = 'game',
	harem = 'harem',
	historical = 'historical',
	horror = 'horror',
	josei = 'josei',
	kids = 'kids',
	martial_arts = 'martial-arts',
	mecha = 'mecha',
	military = 'military',
	music = 'music',
	mystery = 'mystery',
	parody = 'parody',
	police = 'police',
	psychological = 'psychological',
	romance = 'romance',
	samurai = 'samurai',
	school = 'school',
	sci_fi = 'sci-fi',
	seinen = 'seinen',
	shoujo = 'shoujo',
	shoujo_ai = 'shoujo-ai',
	shounen = 'shounen',
	shounen_ai = 'shounen-ai',
	slice_of_life = 'slice-of-life',
	space = 'space',
	sports = 'sports',
	super_power = 'super-power',
	supernatural = 'supernatural',
	thriller = 'thriller',
	vampire = 'vampire',
	yaoi = 'yaoi',
	yuri = 'yuri',
	all = 'all'
}